==> Rscripts/08_plots_analysis.Rmd <==
---
title: "08 Seanuts analysis"
output: 
  html_document: 
    fig_caption: yes
    keep_md: yes
    toc: yes
    theme: "lumen"
    toc_float: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)
```

Load libraries
```{r, message=FALSE, warning=FALSE}
library(stringr)
library(tidyverse)
library(ggplot2)
library(broom)
library(coefplot)
suppressPackageStartupMessages(library(vegan))
library(purrr)
library(arm)
library(ggvegan)
library(dplyr)
```

Read in data 
```{r, message=FALSE, warning=FALSE}
a22 <- read_csv("/Users/Joey/Documents/Nutrient_Analysis/data-processed/seanuts_select_8.csv")
```

### data prep
```{r}
# do a bit of finagling to get it into long form --------------------------

n.long <- a22 %>% 
  dplyr::select(species_name, subgroup, prot_g, protcnt_g, epa, dha, fapun3, ca_mg, fat_g, zn_mg, fe_mg, seanuts_id2, tl, food_item_id_2) %>% 
  gather(key = "nutrient", value = "concentration", prot_g, protcnt_g, epa, dha, fapun3, ca_mg, fat_g, zn_mg, fe_mg) %>% 
  filter(!is.na(concentration)) 
```

```{r}
# begin plots and analysis ------------------------------------------------

g <- ggplot(n.long, aes(concentration)) + geom_histogram(binwidth = 0.07)
g + facet_grid(nutrient ~ ., scales = "free_y") + theme_bw() + scale_x_log10()
```


```{r, message=FALSE, warning=FALSE}
# find the min and max for each nutrient ----------------------------------

n.long %>% 
  group_by(nutrient) %>% 
  distinct(species_name) %>%
  count() %>%
  knitr::kable(align = 'c', format = 'markdown', digits = 2)
```

#### Multivariate stuff

```{r}
# onto multivariate stuff -------------------------------------------------

ntbl.minerals <- a22 %>% 
  dplyr::select(subgroup, species_name, ca_mg, fe_mg, zn_mg)


## ok let's just get rid of the one super outlier ca and fe measurement for now

ntbl.minerals$ca_mg[ntbl.minerals$ca_mg == 41206.00000] <- NA
ntbl.minerals$fe_mg[ntbl.minerals$fe_mg > 40939.00000] <- NA

minerals <- ntbl.minerals
minerals$subgroup <- as.factor(minerals$subgroup)
minerals$species_name <- as.factor(minerals$species_name)

minerals <- minerals %>% 
  filter(!is.na(species_name))

min.mat <- minerals %>% 
  mutate(species_name = as.character(species_name)) %>% 
  group_by(subgroup, species_name) %>% 
  summarise(mean.CA = mean(ca_mg*1000, na.rm = TRUE),
            mean.ZN = mean(zn_mg*1000, na.rm = TRUE), 
            mean.FE = mean(fe_mg*1000, na.rm = TRUE)) %>%
  filter(!is.na(mean.CA), !is.na(mean.ZN), !is.na(mean.FE)) %>%
  ungroup() %>% 
  dplyr::distinct(species_name, .keep_all = TRUE)


matrix.min <- data.matrix(min.mat[, 3:5])
rownames(matrix.min) <- min.mat$species_name 

min.taxon <- minerals %>%
  dplyr::distinct(species_name, subgroup)
  

min.env <- dplyr::semi_join(min.mat, min.taxon, by = "species_name") 
min.env <- min.env %>%
  dplyr::filter(!is.na(species_name)) %>% 
  dplyr::distinct(species_name, .keep_all = TRUE) %>% 
  dplyr::select(subgroup, species_name)

#### begin ordination!

ord.mine <- metaMDS(matrix.min, distance="bray", trymax=200)
ord.mine$stress

## try plotting the ggvegan way

autoplot(ord.mine)
ord_long <- fortify(ord.mine)

### now join the ordination results with the subgroup data

scaling <- left_join(ord_long, min.env, by = c("Label" = "species_name")) %>% 
  filter(Score == "sites")

ggplot(data = scaling, aes(x = Dim1, y = Dim2, colour = subgroup, label = Label)) + geom_point(size = 4) +
  geom_label()


ggplot(data = scaling, aes(x = Dim1, y = Dim2, colour = subgroup)) + geom_point(size = 4) 


site.scaling <- as.data.frame(ord.mine$points)

# site.scaling$nfi_plot <- row.names(site.scaling)
site.scaling$species_name <- row.names(site.scaling)

# min.env$nfi_plot <- row.names(min.env)
# str(min.env)

new.compiled <- full_join(site.scaling, min.env)


plot(ord.mine, type = "n", cex=1) ### looks like Metacarcinus magister is an outlier here
# points(new.compiled$MDS1, new.compiled$MDS2, pch= as.integer(new.compiled$Subgroup), cex = 1)
points(new.compiled$MDS1, new.compiled$MDS2, col = (as.integer(new.compiled$subgroup)), pch= as.integer(new.compiled$subgroup), cex = 1.2)
legend('topleft', legend = levels(new.compiled$subgroup), col = 1:3, pch = 16, cex = 0.8)


ordiplot(ord.mine, type = "text")
ordiellipse(ord.mine, draw = "polygon", new.compiled$subgroup, conf = 0.95, label = T)
ordihull(ord.mine, draw = "polygon", new.compiled$subgroup, conf = 0.95, label = T)
# ordicluster(ord.mine, draw = "polygon", new.compiled$subgroup, conf = 0.95, label = T)
ordispider(ord.mine, new.compiled$subgroup,col="grey")
legend('topleft', legend = levels(new.compiled$subgroup), col = 1:3, pch = 16, cex = 0.8)


### calculate Bray-Curtis distance among samples

comm.bc.dist <- vegdist(matrix.min, method = "bray")
hist(comm.bc.dist)

# cluster communities using average-linkage algorithm
comm.bc.clust <- hclust(comm.bc.dist, method = "average")

# plot cluster diagram
plot(comm.bc.clust, ylab = "Bray-Curtis dissimilarity")

## Use betadisper to test the significance of the multivariate groups
# min.subgroup <- min.env$subgroup
min.subgroup <- scaling$subgroup
length(min.subgroup)


mod <- betadisper(comm.bc.dist, min.subgroup)

## Perform test
anova(mod)

## Permutation test for F
permutest(mod, pairwise = TRUE, permutations = 200)


#### Use adonis to ask whether the group means in multivariate space are different from each other ####

min.subgroup %>% 
  dplyr::data_frame(subgrp = .) %>%
  filter(!is.na(subgrp)) %>%
  adonis(comm.bc.dist ~ subgrp, data = .)
```


#### RDI analysis

```{r}
# RDI analysis ------------------------------------------------------------

aq.wide <- a22 %>% 
  dplyr::select(species_name, subgroup, prot_g, protcnt_g, epa, dha, fapun3, ca_mg, fat_g, zn_mg, fe_mg, seanuts_id2, tl, food_item_id_2)  
  


### how many micronutrient mineral targets does each species reach?
RDI_minerals <- aq.wide %>% 
  group_by(species_name, subgroup) %>% 
  summarise(mean.CA = mean(ca_mg, na.rm = TRUE),
            mean.ZN = mean(zn_mg, na.rm = TRUE), 
            mean.FE = mean(fe_mg, na.rm = TRUE)) %>% 
  mutate(RDI.CA = ifelse(mean.CA > 120, 1, 0)) %>% 
  mutate(RDI.FE = ifelse(mean.FE > 1.8, 1, 0)) %>% 
  mutate(RDI.ZN = ifelse(mean.ZN > 1.1, 1, 0)) %>% 
  ungroup() %>% 
  mutate(RDI.micro.tot = rowSums(.[6:8])) %>% 
  filter(!is.na(RDI.micro.tot)) %>% 
  arrange(., RDI.micro.tot)


RDI_minerals %>% 
  # group_by(species) %>% 
  # mutate(meanRDI = mean(RDI.micro.tot)) %>%
  group_by(subgroup, RDI.micro.tot) %>% 
  summarise(n = n()) %>% 
  mutate(cum.RDI = cumsum(n)) %>%
  ggplot(., aes(x = RDI.micro.tot, y = cum.RDI)) + geom_bar(stat = "identity") + facet_wrap(~ subgroup, scales = "free_y") 

qplot(factor(subgroup), data = RDI_minerals, geom = "bar", fill = factor(RDI.micro.tot)) + theme_bw()

ggplot(RDI_minerals, aes(RDI.micro.tot)) + geom_bar(binwidth = .5) + facet_wrap(~ subgroup, scales = "free_y") 


RDI_minerals %>% 
  dplyr::filter(subgroup == "Molluscs") %>% 
  # distinct(species) %>% 
  dplyr::filter(RDI.micro.tot > 0) %>% 
  dplyr::distinct(species_name) %>% 
  count()


#### species accumulation curves

RDIS <- dplyr::select(RDI_minerals, 6:8)

spa.rand <- specaccum(RDIS, method = "random")
# png(filename = "sac.full.vs.noMoll.png", width = 6, height = 4, units = 'in', res = 300)
spa.rand$sites
plot(spa.rand, col = "cadetblue", lwd = 2, ci = 1, ci.type = "bar", ci.lty = 3,  ci.col = "cadetblue", ylim = c(0,4), xlim = c(0,80), xlab = "number of fish species in diet", ylab = "number of distinct RDI targets reached", main = "25% RDI targets")
abline( v= 15, col = "cadetblue")
abline( v = 26, col = "pink")
# dev.off()



RDI_minerals %>% 
  filter(subgroup == "finfish") %>% 
  dplyr::select(6:8) %>% 
specaccum(., method = "random") %>%
plot(., col = "black", lwd = 4, ci = 1, ci.type = "bar", ci.lty = 3,  ci.col = "blue", ylim = c(0,3.5), xlim = c(0,60), xlab = "number of fish species in diet", ylab = "number of distinct RDI targets reached", main = "micronutrients: 25% RDI targets")


spec <- RDI_minerals %>% 
  filter(subgroup == "finfish") %>% 
  dplyr::select(6:8) %>% 
  specaccum(., method = "random")

str(spec)


spec$richness

sum(spec$richness < 2)


#### now the accumulation curves for the all the nutrients

ntbl.RDI.all <- aq.wide %>% 
  group_by(species_name, subgroup) %>% 
  summarise(mean.CA = mean(ca_mg, na.rm = TRUE),
            mean.ZN = mean(zn_mg, na.rm = TRUE), 
            mean.FE = mean(fe_mg, na.rm = TRUE),
            mean.EPA = mean(epa, na.rm = TRUE),
            mean.DHA = mean(dha, na.rm = TRUE)) %>% 
  mutate(RDI.CA = ifelse(mean.CA > 300, 1, 0)) %>% 
  mutate(RDI.FE = ifelse(mean.FE > 4.5, 1, 0)) %>% 
  mutate(RDI.ZN = ifelse(mean.ZN > 2.75, 1, 0)) %>%
  mutate(RDI.EPA = ifelse(mean.EPA > 0.25, 1, 0)) %>% 
  mutate(RDI.DHA = ifelse(mean.DHA > 0.25, 1, 0)) %>% 
  ungroup() %>% 
  mutate(RDI.micro.tot = rowSums(.[8:12])) %>% 
  filter(!is.na(RDI.micro.tot)) 

### create spa curves for each of the subgroups individually
subgroup_spa <- ntbl.RDI.all %>%
  dplyr::select(-RDI.micro.tot) %>%
  dplyr::select(-contains("mean")) %>% 
  dplyr::select(-species_name) %>% 
  split( .$subgroup) %>% 
  map(.f = `[`, c("RDI.CA", "RDI.FE", "RDI.ZN", "RDI.EPA", "RDI.DHA")) %>%
  map(.f = specaccum, method = "random")


#### how many species do you need to sample before reaching 3 RDI targets?
subgroup_spa %>% 
  map(.f = `[`, "richness") %>% 
  unlist() %>% 
  as.data.frame() %>% 
  knitr::kable()
```

### Traits section 

```{r}
# traits section ----------------------------------------------------------

# seanuts_ecology <- read_csv("data-processed/seanuts_ecology.csv")
seanuts_ecology <- read_csv("/Users/Joey/Documents/Nutrient_Analysis/data-processed/seanuts_select_8.csv")

## ok let's just get rid of the one super outlier ca and fe measurement for now

seanuts_ecology$ca_mg[seanuts_ecology$ca_mg == 41206.00000] <- NA
seanuts_ecology$fe_mg[seanuts_ecology$fe_mg > 40939.00000] <- NA


names_seanuts <- names(seanuts_ecology)
str_subset(names_seanuts, "length")
str_subset(names_seanuts, "Length")
str_subset(names_seanuts, "sl")
str_subset(names_seanuts, "BrackishWater")

sum(!is.na(seanuts_ecology$Length))
sum(!is.na(seanuts_ecology$slmax))
sum(!is.na(seanuts_ecology$length_from_study))

seanuts_ecology <- seanuts_ecology %>% 
  rename(foodtroph = FoodTroph)

n.long <- seanuts_ecology %>% 
  dplyr::select(species_name, subgroup, prot_g, protcnt_g, epa, dha, ca_mg, fat_g,
                zn_mg, fe_mg, slmax, seanuts_id2, tl, food_item_id_2,
                Length, abs_lat, Herbivory2, DemersPelag, contains("Brack"), Marine, Fresh, contains("troph"), contains("length"), contains("weight")) %>% 
  gather(key = "nutrient", value = "concentration", prot_g, protcnt_g, epa, dha, ca_mg, fat_g, zn_mg, fe_mg) %>% 
  filter(!is.na(concentration)) 

ggplot(aes(x = foodtroph, y = tl), data = n.long) + geom_point()
```


```{r}
names(n.long)
unique(n.l)

n.long %>% 
  filter(nutrient %in% c("ca_mg", "fe_mg", "zn_mg")) %>%
  # filter(nutrient %in% c("epa", "dha")) %>%
  # mutate_each_(funs(scale), vars = c("max_length", "TL", "Abs_lat")) %>% 
  group_by(nutrient) %>% 
  do(fit = glance(lm(log(.$concentration) ~ log(CommonLength), data = .), conf.int = TRUE)) %>% 
  unnest(fit) %>% View
  filter(term != "(Intercept)") %>% 
  ggplot(aes(x = term, y = estimate)) + geom_point(size = 2) +
  geom_errorbar(aes(ymin = conf.low, ymax = conf.high), width = 0.1) +
  facet_wrap( ~ nutrient, scales = "free") +
  geom_hline(yintercept = 0) + 
  theme(axis.text.x = element_text(angle = 45, hjust = 1))
  
n.long %>% 
  filter(nutrient == "ca_mg") %>% 
  ggplot(aes(x = log(CommonLength), y = log(concentration))) + geom_point() +
  geom_smooth(method = "lm")


n.long %>% 
  filter(nutrient == "ca_mg") %>% 
  lm(log(concentration) ~ log(Length), data = .) %>% 
  summary
  
```

```{r}
seanuts_ecology %>% 
  # mutate_each_(funs(scale), vars = c("Length", "tl", "abs_lat")) %>% 
  do(fit = tidy(lm(log(prot_g) ~ log(Length) + tl + abs_lat, data = .), conf.int = TRUE)) %>% 
  unnest(fit) %>% 
  filter(term != "(Intercept)") %>%
  ggplot(aes(x = term, y = estimate)) + geom_point(size = 2) +
  geom_errorbar(aes(ymin = conf.low, ymax = conf.high), width = 0.1) +
  geom_hline(yintercept = 0) + 
  theme(axis.text.x = element_text(angle = 45, hjust = 1))
```

```{r}
n.long %>% 
  filter(nutrient == "epa") %>% 
  lm(log(concentration) ~ log(Length) + abs_lat, data = .) %>% 
  summary
```

```{r}

str(n.long)
n.long %>% 
  filter(nutrient == "fe_mg") %>% 
  filter(!is.na(concentration)) %>% 
  lm(concentration ~ log(Length), data = .) %>% 
  summary


n.long %>% 
  filter(nutrient == "fe_mg") %>% 
  ggplot(aes(x = log(Length), y = log(concentration))) + geom_point() +
  geom_smooth(method = "lm")
```
