==> Rscripts/11_nutrient_analysis_figureset.Rmd <==
---
title: "nutrient analysis figure set"
output: 
  html_document: 
    fig_caption: yes
    keep_md: yes
    toc: yes
    toc_float: yes
---

To do list: get the bones no bones incorporated
make table of all the ISSCAAP categories and numbers of distinct species for each nutrient
functional group/morphology/parts eaten
table s1: supplementary table with all the species and the nutrients

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


```{r message=FALSE, warning=FALSE, include=FALSE}
library(stringr)
library(purrr)
library(janitor)
library(MuMIn)
library(broom)
library(forcats)
library(tidyverse)
library(arm)
library(vegan)
library(ggvegan)
library(listviewer)
```

#### Background: dataset summaries

```{r message=FALSE, warning=FALSE}
trait_data <- read_csv("/Users/Joey/Documents/Nutrient_Analysis/data-processed/n.long_lat3.csv")
```

For how many species do we have nutrient data?
```{r}
trait_data %>% 
  group_by(nutrient) %>% 
  distinct(species_name) %>%
  count() %>%
  knitr::kable(align = 'c', format = 'markdown', digits = 2)
```

### Figure 1. Distribution of nutrient concentrations across species
```{r message=FALSE, warning=FALSE}
trait_data <- trait_data %>% 
  mutate(nutrient2 = str_replace(nutrient, "prot_g", "protein_g")) %>% 
  mutate(nutrient2 = str_replace(nutrient2, "protcnt_g", "protein_g")) %>% ### collapsing the various protein measurements into one variable
  rename(nutrient_original = nutrient,
         nutrient = nutrient2)

trait_data %>% 
  mutate(nutrient = str_replace(nutrient, "ca_mg", "calcium")) %>% 
  mutate(nutrient = str_replace(nutrient, "fe_mg", "iron")) %>% 
  mutate(nutrient = str_replace(nutrient, "zn_mg", "zinc")) %>% 
  mutate(nutrient = str_replace(nutrient, "fat_g", "fat")) %>% 
  mutate(nutrient = str_replace(nutrient, "protein_g", "protein")) %>% 
  mutate(nutrient = str_replace(nutrient, "epa", "EPA")) %>% 
  mutate(nutrient = str_replace(nutrient, "dha", "DHA")) %>% 
  filter(!grepl("fapun", nutrient)) %>% 
ggplot(aes(concentration)) + geom_histogram(binwidth = 0.07) +
facet_grid(nutrient ~ ., scales = "free_y") + theme_bw() + scale_x_log10() + xlab("concentration") +
  theme(text = element_text(size=20))
ggsave("/Users/Joey/Documents/Nutrient_Analysis/figures/nutrient_histogram.png", width = 9, height = 10)
  
```

### Figure 2. Few species reach RDI targets for more than one nutrient
```{r}
### how many micronutrient mineral targets does each species reach?
RDI_minerals <- trait_data %>% 
  spread(nutrient, concentration) %>% 
  group_by(species_name, subgroup) %>% 
  summarise(mean.CA = mean(ca_mg, na.rm = TRUE),
            mean.ZN = mean(zn_mg, na.rm = TRUE), 
            mean.FE = mean(fe_mg, na.rm = TRUE)) %>% 
  mutate(RDI.CA = ifelse(mean.CA > 120, 1, 0)) %>% 
  mutate(RDI.FE = ifelse(mean.FE > 1.8, 1, 0)) %>% 
  mutate(RDI.ZN = ifelse(mean.ZN > 1.1, 1, 0)) %>% 
  ungroup() %>% 
  mutate(RDI.micro.tot = rowSums(.[6:8])) %>% 
  filter(!is.na(RDI.micro.tot)) %>% 
  arrange(., RDI.micro.tot)
```

```{r}
RDI_minerals %>% 
  group_by(subgroup, RDI.micro.tot) %>% 
  summarise(n = n()) %>%  
  mutate(cum.RDI = cumsum(n)) %>% 
  ggplot(., aes(x = RDI.micro.tot, y = n)) + geom_bar(stat = "identity") + facet_wrap(~ subgroup, scales = "free_y") + theme_bw() +
  xlab("number of RDI targets in a single 100g portion") +
  ylab("number of species")
```

```{r}
library(ohicore)
RDI_target_table <- RDI_minerals %>% 
  group_by(subgroup, RDI.micro.tot) %>% 
  summarise(n = n()) 

finfish_rdi <- RDI_target_table %>% 
  filter(subgroup == "crustacean")  
PlotFlower(lengths=finfish_rdi$n)
```

#### Multivariate nutrient trait analysis

Get the data in order
```{r warning= FALSE, include = FALSE}
ntbl.minerals <- trait_data %>% 
  spread(nutrient, concentration) %>% 
  dplyr::select(subgroup, species_name, ca_mg, fe_mg, zn_mg)

minerals <- ntbl.minerals
minerals$subgroup <- as.factor(minerals$subgroup)
minerals$species_name <- as.factor(minerals$species_name)

minerals <- minerals %>% 
  filter(!is.na(species_name))

min.mat <- minerals %>% 
  mutate(species_name = as.character(species_name)) %>% 
  group_by(subgroup, species_name) %>% 
  summarise(mean.CA = mean(ca_mg*1000, na.rm = TRUE),
            mean.ZN = mean(zn_mg*1000, na.rm = TRUE), 
            mean.FE = mean(fe_mg*1000, na.rm = TRUE)) %>%
  filter(!is.na(mean.CA), !is.na(mean.ZN), !is.na(mean.FE)) %>%
  ungroup() %>% 
  dplyr::distinct(species_name, .keep_all = TRUE)


matrix.min <- data.matrix(min.mat[, 3:5])
rownames(matrix.min) <- min.mat$species_name 

min.taxon <- minerals %>%
  dplyr::distinct(species_name, subgroup)
  

min.env <- dplyr::semi_join(min.mat, min.taxon, by = "species_name") 
min.env <- min.env %>%
  dplyr::filter(!is.na(species_name)) %>% 
  dplyr::distinct(species_name, .keep_all = TRUE) %>% 
  dplyr::select(subgroup, species_name)
```

Ordination etc

```{r warning = FALSE, message=FALSE, warning=FALSE, include=FALSE}
#### begin ordination!

ord.mine <- metaMDS(matrix.min, distance="bray", trymax=200)
ord.mine$stress
ord_long <- fortify(ord.mine)
```

### Figure 4. Functional groups have distinct multi-nutrient profiles
```{r warning = FALSE, message = FALSE}
scaling <- left_join(ord_long, min.env, by = c("Label" = "species_name")) %>% 
  filter(Score == "sites")

# ggplot(data = scaling, aes(x = Dim1, y = Dim2, colour = subgroup, label = Label)) + geom_point(size = 2) + geom_text() + theme_bw()

ggplot(data = scaling, aes(x = Dim1, y = Dim2, colour = subgroup)) + geom_point(size = 4) +
  theme_bw()
ggsave("/Users/Joey/Documents/Nutrient_Analysis/figures/subgroup_ordination_plot.png")
```

```{r, message=FALSE, warning=FALSE}

### calculate Bray-Curtis distance among samples
comm.bc.dist <- vegdist(matrix.min, method = "bray")
# hist(comm.bc.dist)

# cluster communities using average-linkage algorithm
comm.bc.clust <- hclust(comm.bc.dist, method = "average")

# plot cluster diagram
# plot(comm.bc.clust, ylab = "Bray-Curtis dissimilarity")

## Use betadisper to test the significance of the multivariate groups
# min.subgroup <- min.env$subgroup
min.subgroup <- scaling$subgroup
length(min.subgroup)


mod <- betadisper(comm.bc.dist, min.subgroup)

## Perform test
anova(mod)

## Permutation test for F
permutest(mod, pairwise = TRUE, permutations = 200)


#### Use adonis to ask whether the group means in multivariate space are different from each other ####

min.subgroup %>% 
  dplyr::data_frame(subgrp = .) %>%
  filter(!is.na(subgrp)) %>%
  adonis(comm.bc.dist ~ subgrp, data = .)
```

### Figure 4b. Nutrient-by-nutrient boxplots

```{r, echo=FALSE, message=FALSE, warning=FALSE}
RDIs <- read_csv("/Users/Joey/Documents/Nutrient_Analysis/data-processed/RDIs.csv")
unique(trait_data$nutrient)

RDIs <- RDIs %>% 
  mutate(nutrient = str_to_lower(nutrient)) %>%
  mutate(nutrient = str_replace(nutrient, "fat", "fat_g")) %>% 
  # mutate(nutrient = str_replace(nutrient, "protein", "prot_g")) %>% 
  mutate(nutrient = str_replace(nutrient, "epa_g", "epa")) %>% 
  mutate(nutrient = str_replace(nutrient, "dha_g", "dha")) %>% 
  filter(RDI != 56)

trait_data_prot <- trait_data %>% 
  mutate(nutrient = str_replace(nutrient, "protein_g", "protein")) %>% 
  mutate(nutrient = str_replace(nutrient, "protcnt_g", "protein")) %>% 
  mutate(nutrient = str_replace(nutrient, "prot_g", "protein")) %>% 
  filter(!grepl("^Mohanty", ref_info))

trait_data_rdi <- left_join(trait_data_prot, RDIs, by = "nutrient")

trait_data_rdi <- trait_data_rdi %>% 
  mutate(rdi_percent = (concentration/RDI)*100) 

trait_data_rdi %>% 
  filter(nutrient %in% c("ca_mg", "zn_mg", "fe_mg", "protein", "fat_g", "epa", "dha")) %>% 
  ggplot(., aes(x = subgroup, y = rdi_percent, fill = subgroup, geom = "boxplot")) +
 geom_boxplot() +
  theme_bw() +
  theme(axis.text.x = element_blank()) +
  theme(axis.text=element_text(size=12),
        axis.title=element_text(size=14,face="bold")) +
  geom_hline(yintercept=10) +
  ylab("percentage of RDI in edible portion, log scale") +
  facet_wrap( ~ nutrient, scales = "free") +
  scale_y_log10()

```




### Figure 5. Functional group diversity enhances dietary nutritional diversity  

```{r}
#### now the accumulation curves for the all the nutrients

ntbl.RDI.all <- trait_data %>% 
  spread(nutrient, concentration) %>% 
  group_by(species_name, subgroup) %>% 
  summarise(mean.CA = mean(ca_mg, na.rm = TRUE),
            mean.ZN = mean(zn_mg, na.rm = TRUE), 
            mean.FE = mean(fe_mg, na.rm = TRUE),
            mean.EPA = mean(epa, na.rm = TRUE),
            mean.DHA = mean(dha, na.rm = TRUE)) %>% 
  mutate(RDI.CA = ifelse(mean.CA > 300, 1, 0)) %>% 
  mutate(RDI.FE = ifelse(mean.FE > 4.5, 1, 0)) %>% 
  mutate(RDI.ZN = ifelse(mean.ZN > 2.75, 1, 0)) %>%
  mutate(RDI.EPA = ifelse(mean.EPA > 0.25, 1, 0)) %>% 
  mutate(RDI.DHA = ifelse(mean.DHA > 0.25, 1, 0)) %>% 
  ungroup() %>% 
  mutate(RDI.micro.tot = rowSums(.[8:12])) %>% 
  filter(!is.na(RDI.micro.tot)) 

RDI_minerals <- trait_data %>% 
  spread(nutrient, concentration) %>% 
  group_by(species_name, subgroup) %>% 
  summarise(mean.CA = mean(ca_mg, na.rm = TRUE),
            mean.ZN = mean(zn_mg, na.rm = TRUE), 
            mean.FE = mean(fe_mg, na.rm = TRUE)) %>% 
  mutate(RDI.CA = ifelse(mean.CA > 120, 1, 0)) %>% 
  mutate(RDI.FE = ifelse(mean.FE > 1.8, 1, 0)) %>% 
  mutate(RDI.ZN = ifelse(mean.ZN > 1.1, 1, 0)) %>% 
  ungroup() %>% 
  mutate(RDI.micro.tot = rowSums(.[6:8])) %>% 
  filter(!is.na(RDI.micro.tot)) %>% 
  arrange(., RDI.micro.tot)

```

```{r}
#### how many species do you need to sample before reaching 3 RDI targets?
### create spa curves for each of the subgroups individually
subgroup_spa <- ntbl.RDI.all %>%
  dplyr::select(-RDI.micro.tot) %>%
  dplyr::select(-contains("mean")) %>% 
  dplyr::select(-species_name) %>%
  split( .$subgroup) %>% 
  map(.f = `[`, c("RDI.CA", "RDI.FE", "RDI.ZN", "RDI.EPA", "RDI.DHA")) %>%
  map(.f = specaccum, method = "random")


all_spa <- ntbl.RDI.all %>%
  dplyr::select(-RDI.micro.tot) %>%
  dplyr::select(-contains("mean")) %>% 
  dplyr::select(-species_name) %>%
  mutate(subgroup = "all") %>% 
  split( .$subgroup) %>% 
  map(.f = `[`, c("RDI.CA", "RDI.FE", "RDI.ZN", "RDI.EPA", "RDI.DHA")) %>%
  map(.f = specaccum, method = "random")



subgroup_spa <- RDI_minerals %>%
  dplyr::select(-RDI.micro.tot) %>%
  dplyr::select(-contains("mean")) %>%
  dplyr::select(-species_name) %>% 
  split( .$subgroup) %>% 
  map(.f = `[`, c("RDI.CA", "RDI.FE", "RDI.ZN")) %>%
  map(.f = specaccum, method = "random", permutations = 10000)

subgroup_spa_all <- RDI_minerals %>%
  dplyr::select(-RDI.micro.tot) %>%
  dplyr::select(-contains("mean")) %>%
  dplyr::select(-species_name) %>%
    dplyr::select(-subgroup) %>%
 specaccum(method = "random", permutations = 100)

subgroup_spa_all_no_mollusc <- RDI_minerals %>%
  filter(subgroup == "finfish") %>% 
  dplyr::select(-RDI.micro.tot) %>%
  dplyr::select(-contains("mean")) %>%
  dplyr::select(-species_name) %>%
    dplyr::select(-subgroup) %>%
 specaccum(method = "random", permutations =100)


nutraccum <- data.frame(subgroup_spa_all$richness, subgroup_spa_all$site, subgroup_spa_all$sd)
nutraccum_no_mollusc <- data.frame(subgroup_spa_all_no_mollusc$richness, subgroup_spa_all_no_mollusc$site, subgroup_spa_all_no_mollusc$sd)

nutraccum_no_mollusc$groups <- "no mollusc"
nutraccum$groups <- "all"
names(nutraccum_no_mollusc)

nutraccum_no_mollusc <- nutraccum_no_mollusc %>% 
  rename(richness = subgroup_spa_all_no_mollusc.richness,
         site = subgroup_spa_all_no_mollusc.site, 
         sd = subgroup_spa_all_no_mollusc.sd)

nutraccum <- nutraccum %>% 
  rename(richness = subgroup_spa_all.richness, 
         site = subgroup_spa_all.site,
         sd = subgroup_spa_all.sd)
  
nutraccum_all <- bind_rows(nutraccum, nutraccum_no_mollusc)

ggplot(data = nutraccum_all, aes(x = site, y = richness, color = groups)) + geom_line(size = 2) + xlim(0,10) + 
  # geom_ribbon(aes(ymin = richness - sd, ymax = richness + sd, fill = groups), alpha = 0.5) +
  theme_bw()


mod1 <- nls(richness ~ SSlogis(site, Asym, xmid, scale), data = nutraccum)
coef(summary(mod1))

mod2 <- nls(richness ~ SSlogis(site, Asym, xmid, scale), data = nutraccum_no_mollusc)
coef(summary(mod2))

(tidy(mod1, conf.int = TRUE))

accumulated_targets <- subgroup_spa %>% 
  map(.f = `[`, "richness") %>% 
  unlist() %>% 
  as.data.frame()

accumulated_targets_sd <- subgroup_spa %>% 
  map(.f = `[`, "sd") %>% 
  unlist() %>% 
  as.data.frame()

accumulated_targets$richness_level = rownames(accumulated_targets)
colnames(accumulated_targets) <- c("number_of_targets", "richness_level")

accumulated_targets_sd$sd = rownames(accumulated_targets_sd)
colnames(accumulated_targets_sd) <- c("sd", "number_of_targets")

accumulated_targets_sd <- accumulated_targets_sd %>% 
  separate(number_of_targets, into = c("subgroup", "number_of_species")) %>%
  mutate(number_of_species = str_replace(number_of_species, "sd", "")) %>%
  mutate(number_of_species = as.numeric(number_of_species))


accumulated_targets <- accumulated_targets %>% 
  separate(richness_level, into = c("subgroup", "number_of_species")) %>%
  mutate(number_of_species = str_replace(number_of_species, "richness", "")) %>%
  mutate(number_of_species = as.numeric(number_of_species))

accumulated_targets_all <- left_join(accumulated_targets, accumulated_targets_sd)
accumulated_targets_all <- accumulated_targets_all %>% 
  mutate(se = sd / sqrt(number_of_species)) 

## add in the all accumulation curve

accumulated_targets_full <- all_spa %>% 
  map(.f = `[`, "richness") %>% 
  unlist() %>% 
  as.data.frame()

accumulated_targets_full_sd <- all_spa %>% 
  map(.f = `[`, "sd") %>% 
  unlist() %>% 
  as.data.frame()

accumulated_targets_full$richness_level = rownames(accumulated_targets_full)
colnames(accumulated_targets_full) <- c("number_of_targets", "richness_level")

accumulated_targets_full_sd$sd = rownames(accumulated_targets_full_sd)
colnames(accumulated_targets_full_sd) <- c("sd", "number_of_targets")

accumulated_targets_full_sd <- accumulated_targets_full_sd %>% 
  separate(number_of_targets, into = c("subgroup", "number_of_species")) %>%
  mutate(number_of_species = str_replace(number_of_species, "sd", "")) %>%
  mutate(number_of_species = as.numeric(number_of_species))


accumulated_targets_full <- accumulated_targets_full %>% 
  separate(richness_level, into = c("subgroup", "number_of_species")) %>%
  mutate(number_of_species = str_replace(number_of_species, "richness", "")) %>%
  mutate(number_of_species = as.numeric(number_of_species))

accumulated_targets_all_full <- left_join(accumulated_targets_full, accumulated_targets_full_sd)
accumulated_targets_all_full <- accumulated_targets_all_full %>% 
  mutate(se = sd / sqrt(number_of_species)) 

## now bind rows with the subroup split accum

all <- bind_rows(accumulated_targets_all, accumulated_targets_all_full)

all %>%
  filter(number_of_species < 15) %>% 
  rename(group = subgroup) %>% 
ggplot(data = ., aes(x = number_of_species, y = number_of_targets, color = group)) + geom_line(size =1, linetype = "dotted") + theme_bw() + geom_ribbon(aes(ymin = number_of_targets - se, ymax = number_of_targets + se), alpha = 0.2, size = 1) + ylab("number of nutrient requirements fulfilled") +
  xlab("number of species") + theme(text = element_text(size=14)) 

ggsave("/Users/Joey/Documents/Nutrient_Analysis/figures/nutrient_accum_grey.png", width = 8, height = 6)
ggsave("/Users/Joey/Documents/Nutrient_Analysis/figures/nutrient_accum_all.png", width = 8, height = 6)
```
```{r}

# ntbl.RDI.all %>% 
#   dplyr::select(8:12) %>% 
# specaccum(., method = "random") %>%
# plot(., col = "black", lwd = 4, ci = 1, ci.type = "bar", ci.lty = 3,  ci.col = "blue", ylim = c(0,6), xlab = "number of fish species in diet", ylab = "number of distinct RDI targets reached", main = "micronutrients: 25% RDI targets")

RDIS <- dplyr::select(ntbl.RDI.all, 8:12)

spa.rand <- specaccum(RDIS, method = "random")
# plot(spa.rand, col = "cadetblue", lwd = 2, ci = 1, ci.type = "bar", ci.lty = 3,  ci.col = "cadetblue", ylim = c(0,6), xlim = c(0,80), xlab = "number of fish species in diet", ylab = "number of distinct RDI targets reached", main = "25% RDI targets")
# abline( v= 15, col = "cadetblue")
# abline( v = 26, col = "pink")



```


### Figure 5b. What happens when we lose the molluscs -- does the slope of the accumulation curve increase?
```{r}
nutraccum_all %>% 
  group_by(groups) %>% 
  do(tidy(nls(richness ~ SSlogis(site, asymptote, xmid, scale), data = .), conf.int = TRUE)) %>% 
  ggplot(data = ., aes(x = term, y = estimate, color = groups)) + geom_point(size = 2) +
  geom_errorbar(aes(ymin = conf.low, ymax = conf.high), width = 0.1) + theme_bw()

```




```{r, eval=FALSE, include=FALSE}
RDI_mollusc <- RDI_minerals %>%
  dplyr::select(-RDI.micro.tot) %>%
  dplyr::select(-contains("mean")) %>%
  dplyr::select(-species_name) %>% 
  filter(subgroup == "mollusc") %>% 
  dplyr::select(-subgroup)

RDI_fish <- RDI_minerals %>%
  dplyr::select(-RDI.micro.tot) %>%
  dplyr::select(-contains("mean")) %>%
  dplyr::select(-species_name) %>% 
  filter(subgroup == "finfish") %>% 
  dplyr::select(-subgroup)

RDI_no_mollusc <- RDI_minerals %>%
  dplyr::select(-RDI.micro.tot) %>%
  dplyr::select(-contains("mean")) %>%
  dplyr::select(-species_name) %>% 
  filter(subgroup != "mollusc") %>% 
  dplyr::select(-subgroup)


RDI_all <- RDI_minerals %>%
  dplyr::select(-RDI.micro.tot) %>%
  dplyr::select(-contains("mean")) %>%
  dplyr::select(-species_name) %>% 
  dplyr::select(-subgroup)

spa_rand_all <- specaccum(RDI_all, "rarefaction")
mod_all <- fitspecaccum(spa_rand_all, model = "gleason")


spa_rand_fish <- specaccum(RDI_fish, "rarefaction")
mod_fish <- fitspecaccum(spa_rand_fish, model = "gleason")

spa_rand_mollusc <- specaccum(RDI_mollusc, "rarefaction")
mod_mollusc <- fitspecaccum(spa_rand_mollusc, model = "gleason")

spa_rand_no_mollusc <- specaccum(RDI_no_mollusc, method = "random", permutations = 100)
mod_no_mollusc <- fitspecaccum(spa_rand_no_mollusc, model = "michaelis-menten")

# as.data.frame(spa_rand_no_mollusc$perm) %>% 
#   rowwise() %>% 
#   mutate(avg_targets = mean(1:100))

coef(mod_mollusc)
summary(spa_rand_no_mollusc)
coef(mod_fish)
coef(mod_all)
coef(summary(spa_rand_no_mollusc))

confint(mod_fish)

specslope(spa_rand_all, at = 3) 
specslope(spa_rand_fish, at = 3) 
specslope(spa_rand_mollusc, at = 3) 
```

#### Trait analysis

```{r message=FALSE, warning=FALSE}

mod_all <- trait_data %>% 
  filter(concentration > 0) %>% 
  mutate(anacat = ifelse(subgroup != "finfish", "non-migratory", anacat)) %>% 
  filter(!is.na(bulk_max_length), !is.na(bulk_trophic_level), !is.na(feeding_level), !is.na(feeding_mode), !is.na(abs_lat)) %>% 
  mutate(log_length = log(bulk_max_length),
         log_concentration = log(concentration)) %>% 
  filter(!grepl("^Mohanty, B. P.,", ref_info))
```

#### Scatterplots
### Figure 6. Micronutrient concentration decreases with increasing body size
```{r, message=FALSE, warning=FALSE}
trait_data %>% 
  # filter(nutrient == "ca_mg") %>% 
  filter(concentration > 0) %>%
  filter(nutrient %in% c("fe_mg", "ca_mg", "zn_mg")) %>% 
    mutate(nutrient = str_replace(nutrient, "ca_mg", "calcium")) %>% 
  mutate(nutrient = str_replace(nutrient, "fe_mg", "iron")) %>% 
  mutate(nutrient = str_replace(nutrient, "zn_mg", "zinc")) %>%
  group_by(nutrient) %>% 
ggplot(aes(y = log(concentration), x = log(bulk_mean_length)), data = .) + geom_point(size = 3, alpha = 0.5) + geom_smooth(method = "lm", color = "grey") +
  theme_bw() + ylab("ln concentration (mg/100g)") + xlab("ln body length") + facet_wrap( ~ nutrient, scales = "free") +
  theme(text = element_text(size=16)) 
ggsave("/Users/Joey/Documents/Nutrient_Analysis/figures/concentration_body_size.png", width = 8, height = 3)
```


```{r message=FALSE, warning=FALSE}
trait_data %>% 
  # filter(nutrient == "ca_mg") %>% 
  filter(concentration > 0) %>%
  filter(nutrient %in% c("fe_mg", "ca_mg", "zn_mg")) %>% 
  group_by(nutrient) %>% 
  ggplot(aes(y = log(concentration), x = abs_lat), data = .) + geom_point(size = 3, alpha = 0.5, color = "blue") + geom_smooth(method = "lm") +
  theme_bw() + ylab("ln nutrient concentration") + xlab("absolute latitude") + facet_wrap( ~ nutrient, scales = "free")
```

```{r message=FALSE, warning=FALSE}
trait_data %>% 
  # filter(nutrient == "ca_mg") %>% 
  filter(concentration > 0) %>%
  filter(nutrient %in% c("fe_mg", "ca_mg", "zn_mg")) %>% 
  group_by(nutrient) %>% 
  ggplot(aes(y = log(bulk_mean_length), x = abs_lat), data = .) + geom_point(size = 3, alpha = 0.5, color = "blue") + geom_smooth(method = "lm") +
  theme_bw() + ylab("ln body length") + xlab("absolute latitude") + facet_wrap( ~ nutrient, scales = "free")
```


```{r, message=FALSE, warning=FALSE}
## prep the data
mod_all <- trait_data %>% 
  filter(concentration > 0) %>% 
  mutate(anacat = ifelse(subgroup != "finfish", "non-migratory", anacat)) %>% 
  filter(!is.na(bulk_max_length), !is.na(bulk_trophic_level), !is.na(feeding_level), !is.na(feeding_mode), !is.na(abs_lat)) %>% 
  mutate(log_length = log(bulk_max_length),
         log_concentration = log(concentration)) %>% 
  filter(!grepl("^Mohanty, B. P.,", ref_info))
```


Iron
```{r echo = FALSE}
mod <- mod_all %>% 
  filter(nutrient == "fe_mg") %>% 
  filter(subgroup == "finfish")

mod1 <- standardize(lm(log_concentration ~ log_length + bulk_trophic_level + feeding_mode + abs_lat, data = mod), standardize.y = TRUE) 
mod2 <- standardize(lm(log_concentration ~ log_length + bulk_trophic_level + feeding_level + abs_lat, data = mod), standardize.y = TRUE) 
mod3 <- standardize(lm(log_concentration ~ log_length + feeding_level + feeding_mode + abs_lat, data = mod), standardize.y = TRUE) 
mod4 <- standardize(lm(log_concentration ~ bulk_trophic_level + feeding_mode + abs_lat, data = mod), standardize.y = TRUE) 
mod5 <- standardize(lm(log_concentration ~ log_length + feeding_level + feeding_mode + abs_lat, data = mod), standardize.y = TRUE) 
mod6 <- standardize(lm(log_concentration ~ log_length + bulk_trophic_level + feeding_level, data = mod), standardize.y = TRUE) 
mod7 <- standardize(lm(log_concentration ~ bulk_trophic_level + feeding_level + feeding_mode + abs_lat, data = mod), standardize.y = TRUE) 

ddfe <- model.sel(mod1, mod2, mod3, mod4, mod5, mod6, mod7)
ddfe
summary(mod2)
fe_CI_average <- rownames_to_column(as.data.frame(confint(model.avg(ddfe, subset = cumsum(weight) <= .95)), var = "term")) %>% rename(conf_low = `2.5 %`,
         conf_high = `97.5 %`) %>% 
  rename(term = rowname)
fe_slopes_average <- enframe(coef(model.avg(ddfe, subset = cumsum(weight) <= .95)), name = "term", value = "slope")
fe_results <- left_join(fe_CI_average, fe_slopes_average, by = "term") %>% 
  mutate(nutrient = "iron")

summary(mod1)


```

Zinc
```{r echo=FALSE}
mod <- mod_all %>% 
  filter(nutrient == "zn_mg") %>% 
  filter(subgroup == "finfish")

mod1 <- standardize(lm(log_concentration ~ log_length + bulk_trophic_level + feeding_mode + abs_lat, data = mod), standardize.y = TRUE) 
mod2 <- standardize(lm(log_concentration ~ log_length + bulk_trophic_level + feeding_level + abs_lat, data = mod), standardize.y = TRUE) 
mod3 <- standardize(lm(log_concentration ~ log_length + feeding_level + feeding_mode + abs_lat, data = mod), standardize.y = TRUE) 
mod4 <- standardize(lm(log_concentration ~ bulk_trophic_level + feeding_mode + abs_lat, data = mod), standardize.y = TRUE) 
mod5 <- standardize(lm(log_concentration ~ log_length + feeding_level + feeding_mode + abs_lat, data = mod), standardize.y = TRUE) 
mod6 <- standardize(lm(log_concentration ~ log_length + bulk_trophic_level + feeding_level, data = mod), standardize.y = TRUE) 
mod7 <- standardize(lm(log_concentration ~ bulk_trophic_level + feeding_level + feeding_mode + abs_lat, data = mod), standardize.y = TRUE) 

ddzn <- model.sel(mod1, mod2, mod3, mod4, mod5, mod6, mod7)
summary(mod1)
zn_CI_average <- rownames_to_column(as.data.frame(confint(model.avg(ddzn, subset = cumsum(weight) <= .95)), var = "term")) %>% rename(conf_low = `2.5 %`,
         conf_high = `97.5 %`) %>% 
  rename(term = rowname)
zn_slopes_average <- enframe(coef(model.avg(ddzn, subset = cumsum(weight) <= .95)), name = "term", value = "slope")
zn_results <- left_join(zn_CI_average, zn_slopes_average, by = "term") %>% 
  mutate(nutrient = "zinc")
```


Calcium
```{r echo=FALSE, message=FALSE, warning=FALSE}
mod <- mod_all %>% 
  filter(nutrient == "ca_mg") %>% 
  filter(subgroup == "finfish")

mod1 <- standardize(lm(log_concentration ~ log_length + bulk_trophic_level + feeding_mode + abs_lat, data = mod), standardize.y = TRUE) 
mod2 <- standardize(lm(log_concentration ~ log_length + bulk_trophic_level + feeding_level + abs_lat, data = mod), standardize.y = TRUE) 
mod3 <- standardize(lm(log_concentration ~ log_length + feeding_level + feeding_mode + abs_lat, data = mod), standardize.y = TRUE) 
mod4 <- standardize(lm(log_concentration ~ bulk_trophic_level + feeding_mode + abs_lat, data = mod), standardize.y = TRUE) 
mod5 <- standardize(lm(log_concentration ~ log_length + feeding_level + feeding_mode + abs_lat, data = mod), standardize.y = TRUE) 
mod6 <- standardize(lm(log_concentration ~ log_length + bulk_trophic_level + feeding_level, data = mod), standardize.y = TRUE) 
mod7 <- standardize(lm(log_concentration ~ bulk_trophic_level + feeding_level + feeding_mode + abs_lat, data = mod), standardize.y = TRUE) 

ddca <- model.sel(mod1, mod2, mod3, mod4, mod5, mod6, mod7)
summary(mod2)
ca_CI_average <- rownames_to_column(as.data.frame(confint(model.avg(ddca, subset = cumsum(weight) <= .95)), var = "term")) %>% rename(conf_low = `2.5 %`,
         conf_high = `97.5 %`) %>% 
  rename(term = rowname)
ca_slopes_average <- enframe(coef(model.avg(ddca, subset = cumsum(weight) <= .95)), name = "term", value = "slope")
ca_results <- left_join(ca_CI_average, ca_slopes_average, by = "term") %>% 
  mutate(nutrient = "calcium")
```

EPA
```{r echo=FALSE}
mod <- mod_all %>% 
  filter(nutrient == "epa") %>% 
  filter(subgroup == "finfish")

mod1 <- standardize(lm(log_concentration ~ log_length + bulk_trophic_level + feeding_mode + abs_lat, data = mod), standardize.y = TRUE) 
mod2 <- standardize(lm(log_concentration ~ log_length + bulk_trophic_level + feeding_level + abs_lat, data = mod), standardize.y = TRUE) 
mod3 <- standardize(lm(log_concentration ~ log_length + feeding_level + feeding_mode + abs_lat, data = mod), standardize.y = TRUE) 
mod4 <- standardize(lm(log_concentration ~ bulk_trophic_level + feeding_mode + abs_lat, data = mod), standardize.y = TRUE) 
mod5 <- standardize(lm(log_concentration ~ log_length + feeding_level + feeding_mode + abs_lat, data = mod), standardize.y = TRUE) 
mod6 <- standardize(lm(log_concentration ~ log_length + bulk_trophic_level + feeding_level, data = mod), standardize.y = TRUE) 
mod7 <- standardize(lm(log_concentration ~ bulk_trophic_level + feeding_level + feeding_mode + abs_lat, data = mod), standardize.y = TRUE) 

ddepa <- model.sel(mod1, mod2, mod3, mod4, mod5, mod6, mod7)
summary(mod1)
epa_CI_average <- rownames_to_column(as.data.frame(confint(model.avg(ddepa, subset = cumsum(weight) <= .95)), var = "term")) %>% rename(conf_low = `2.5 %`,
         conf_high = `97.5 %`) %>% 
  rename(term = rowname)
epa_slopes_average <- enframe(coef(model.avg(ddepa, subset = cumsum(weight) <= .95)), name = "term", value = "slope")
epa_results <- left_join(epa_CI_average, epa_slopes_average, by = "term") %>% 
  mutate(nutrient = "epa")
```

DHA
```{r echo=FALSE}
mod <- mod_all %>% 
  filter(nutrient == "dha") %>% 
  filter(subgroup == "finfish")

mod1 <- standardize(lm(log_concentration ~ log_length + bulk_trophic_level + feeding_mode + abs_lat, data = mod), standardize.y = TRUE) 
mod2 <- standardize(lm(log_concentration ~ log_length + bulk_trophic_level + feeding_level + abs_lat, data = mod), standardize.y = TRUE) 
mod3 <- standardize(lm(log_concentration ~ log_length + feeding_level + feeding_mode + abs_lat, data = mod), standardize.y = TRUE) 
mod4 <- standardize(lm(log_concentration ~ bulk_trophic_level + feeding_mode + abs_lat, data = mod), standardize.y = TRUE) 
mod5 <- standardize(lm(log_concentration ~ log_length + feeding_level + feeding_mode + abs_lat, data = mod), standardize.y = TRUE) 
mod6 <- standardize(lm(log_concentration ~ log_length + bulk_trophic_level + feeding_level, data = mod), standardize.y = TRUE) 
mod7 <- standardize(lm(log_concentration ~ bulk_trophic_level + feeding_level + feeding_mode + abs_lat, data = mod), standardize.y = TRUE) 

dddha <- model.sel(mod1, mod2, mod3, mod4, mod5, mod6, mod7)
summary(mod6)
dha_CI_average <- rownames_to_column(as.data.frame(confint(model.avg(dddha, subset = cumsum(weight) <= .95)), var = "term")) %>% rename(conf_low = `2.5 %`,
         conf_high = `97.5 %`) %>% 
  rename(term = rowname)
dha_slopes_average <- enframe(coef(model.avg(dddha, subset = cumsum(weight) <= .95)), name = "term", value = "slope")
dha_results <- left_join(dha_CI_average, dha_slopes_average, by = "term") %>% 
  mutate(nutrient = "dha")
```

fat
```{r echo=FALSE}
mod <- mod_all %>% 
  filter(nutrient == "fat_g") %>% 
  filter(subgroup == "finfish")

mod1 <- standardize(lm(log_concentration ~ log_length + bulk_trophic_level + feeding_mode + abs_lat, data = mod), standardize.y = TRUE) 
mod2 <- standardize(lm(log_concentration ~ log_length + bulk_trophic_level + feeding_level + abs_lat, data = mod), standardize.y = TRUE) 
mod3 <- standardize(lm(log_concentration ~ log_length + feeding_level + feeding_mode + abs_lat, data = mod), standardize.y = TRUE) 
mod4 <- standardize(lm(log_concentration ~ bulk_trophic_level + feeding_mode + abs_lat, data = mod), standardize.y = TRUE) 
mod5 <- standardize(lm(log_concentration ~ log_length + feeding_level + feeding_mode + abs_lat, data = mod), standardize.y = TRUE) 
mod6 <- standardize(lm(log_concentration ~ log_length + bulk_trophic_level + feeding_level, data = mod), standardize.y = TRUE) 
mod7 <- standardize(lm(log_concentration ~ bulk_trophic_level + feeding_level + feeding_mode + abs_lat, data = mod), standardize.y = TRUE) 

ddfat <- model.sel(mod1, mod2, mod3, mod4, mod5, mod6, mod7)
summary(mod1)
summary(model.avg(ddfat, subset = cumsum(weight) <= .95))

fat_CI_average <- rownames_to_column(as.data.frame(confint(model.avg(ddfat, subset = cumsum(weight) <= .95)), var = "term")) %>% rename(conf_low = `2.5 %`,
         conf_high = `97.5 %`) %>% 
  rename(term = rowname)
fat_slopes_average <- enframe(coef(model.avg(ddfat, subset = cumsum(weight) <= .95)), name = "term", value = "slope")
fat_results <- left_join(fat_CI_average, fat_slopes_average, by = "term") %>% 
  mutate(nutrient = "fat")
```


protein
```{r echo=FALSE}
unique(mod_all$nutrient)

mod <- mod_all %>% 
  filter(nutrient %in% c("protcnt_g")) %>% 
  filter(subgroup == "finfish")

mod1 <- standardize(lm(log_concentration ~ log_length + bulk_trophic_level + feeding_mode + abs_lat, data = mod), standardize.y = TRUE) 
mod2 <- standardize(lm(log_concentration ~ log_length + bulk_trophic_level + feeding_level + abs_lat, data = mod), standardize.y = TRUE) 
mod3 <- standardize(lm(log_concentration ~ log_length + feeding_level + feeding_mode + abs_lat, data = mod), standardize.y = TRUE) 
mod4 <- standardize(lm(log_concentration ~ bulk_trophic_level + feeding_mode + abs_lat, data = mod), standardize.y = TRUE) 
mod5 <- standardize(lm(log_concentration ~ log_length + feeding_level + feeding_mode + abs_lat, data = mod), standardize.y = TRUE) 
mod6 <- standardize(lm(log_concentration ~ log_length + bulk_trophic_level + feeding_level, data = mod), standardize.y = TRUE) 
mod7 <- standardize(lm(log_concentration ~ bulk_trophic_level + feeding_level + feeding_mode + abs_lat, data = mod), standardize.y = TRUE) 

ddprotein <- model.sel(mod1, mod2, mod3, mod4, mod5, mod6, mod7)
summary(mod1)
protein_CI_average <- rownames_to_column(as.data.frame(confint(model.avg(ddprotein, subset = cumsum(weight) <= .95)), var = "term")) %>% rename(conf_low = `2.5 %`,
         conf_high = `97.5 %`) %>% 
  rename(term = rowname)
protein_slopes_average <- enframe(coef(model.avg(ddprotein, subset = cumsum(weight) <= .95)), name = "term", value = "slope")
protein_results <- left_join(protein_CI_average, protein_slopes_average, by = "term") %>% 
  mutate(nutrient = "protein")
```


### Figure 7. Within functional groups, some traits such body size and latitude are strongly associated with nutritional profile 
```{r echo=FALSE}
microelements_finfish <- bind_rows(fe_results, zn_results, ca_results, epa_results, dha_results, fat_results, protein_results) %>% 
  filter(term != "(Intercept)") %>% 
  mutate(term = str_replace(term, "z.log_length", "body size (log length)")) %>% 
  mutate(term = str_replace(term, "z.bulk_trophic_level", "fractional trophic position")) %>% 
  mutate(term = str_replace(term, "feeding_modefiltering plankton", "plankton feeder")) %>% 
  mutate(term = str_replace(term, "feeding_modegrazing on aquatic plants", "herbivore grazer")) %>% 
  mutate(term = str_replace(term, "feeding_modeselective plankton feeding", "selective filter feeder")) %>%
  mutate(term = str_replace(term, "feeding_modevariable", "variable feeding mode")) %>%
  mutate(term = str_replace(term, "z.abs_lat", "absolute latitude")) %>%
  mutate(term = ifelse(term == "feeding_levelplants/detritus+animals (troph. 2.2-2.79)", "omnivore", term)) %>%
  mutate(term = ifelse(term == "feeding_levelmainly animals (troph. 2.8 and up)", "carnivore", term)) %>% 
  mutate(term = ifelse(term == "feeding_modehunting macrofauna (predator)", "active predator", term)) %>% 
  mutate(term = ifelse(term == "feeding_levelmainly plants/detritus (troph. 2-2.19)", "herbivore", term)) %>% 
  mutate(term = ifelse(term == "feeding_modeother", "other feeding mode", term)) %>% 
  mutate(term = ifelse(term == "feeding_modefeeding on the prey of a host (commensal)", "commensal", term)) 

microelements %>% 
  filter(nutrient %in% c("calcium", "iron", "zinc")) %>% 
ggplot(data = ., aes(x = term, y = slope)) + geom_point(size = 4) + 
  geom_hline(yintercept = 0) +
  geom_errorbar(aes(ymin = conf_low, ymax = conf_high), width = 0.1) +
  coord_flip() + theme_bw() + facet_wrap( ~ nutrient) +
  theme(text = element_text(size=12)) +
  ylab("model-averaged estimate") 
ggsave("/Users/Joey/Documents/Nutrient_Analysis/figures/coefs_dot_plot.png", width = 8, height = 5)
```

```{r}
microelements %>% 
  filter(nutrient %in% c("fat", "protein", "epa", "dha")) %>% 
ggplot(data = ., aes(x = term, y = slope)) + geom_point(size = 4) + 
  geom_hline(yintercept = 0) +
  geom_errorbar(aes(ymin = conf_low, ymax = conf_high), width = 0.1) +
  coord_flip() + theme_bw() + facet_wrap( ~ nutrient) +
  theme(text = element_text(size=12)) +
  ylab("model-averaged estimate") 
ggsave("/Users/Joey/Documents/Nutrient_Analysis/figures/coefs_dot_plot.png", width = 8, height = 5)


```


CINE body part analysis
```{r}
CINE_merge <- read_csv("/Users/Joey/Documents/Nutrient_Analysis/data-processed/CINE-body-parts.csv")


ca_plot <- CINE_merge %>% 
  filter(nutrient %in% c("ca_mg", "zn_mg", "fe_mg", "mn_mg")) %>% 
  filter(part != "not specified") %>% 
  filter(nutrient == "ca_mg") %>% 
  filter(part != "tongues + cheeks") %>% 
  filter(part != "whole, no skin") %>% 
  mutate(nutrient = str_replace(nutrient, "ca_mg", "calcium")) %>% 
  group_by(nutrient, part) %>% 
  summarise_each(funs(mean, std.error), concentration) %>%
  group_by(nutrient) %>% 
  arrange(mean) %>% 
  ggplot(aes(x = reorder(part, mean), y = mean)) + geom_point(size = 3) +
  geom_errorbar(aes(ymin = mean - std.error, ymax = mean + std.error), width = 0.2) +
  facet_wrap( ~ nutrient, scales = "free")+
  geom_hline(yintercept = 1200/10) +
  theme_bw() +
  theme(text = element_text(size=20),
        axis.text.x = element_text(angle=45, hjust=1)) + xlab("body part") + ylab("nutrient concentration (mg/100g edible portion)")


fe_plot <- CINE_merge %>% 
  filter(nutrient %in% c("ca_mg", "zn_mg", "fe_mg", "mn_mg")) %>% 
  filter(part != "not specified") %>% 
  filter(nutrient == "fe_mg") %>% 
  filter(part != "tongues + cheeks") %>% 
  filter(part != "whole, no skin") %>% 
  mutate(nutrient = str_replace(nutrient, "fe_mg", "iron")) %>% 
  group_by(nutrient, part) %>% 
  summarise_each(funs(mean, std.error), concentration) %>%
  group_by(nutrient) %>% 
  arrange(mean) %>% 
  ggplot(aes(x = reorder(part, mean), y = mean)) + geom_point(size = 3) +
  geom_errorbar(aes(ymin = mean - std.error, ymax = mean + std.error), width = 0.2) +
  facet_wrap( ~ nutrient, scales = "free")+
  geom_hline(yintercept = 18/10) +
  theme_bw() +
  theme(text = element_text(size=20),
        axis.text.x = element_text(angle=45, hjust=1)) + xlab("body part") + ylab("nutrient concentration (mg/100g edible portion)")

zn_plot <- CINE_merge %>% 
  filter(nutrient %in% c("ca_mg", "zn_mg", "fe_mg", "mn_mg")) %>% 
  filter(part != "not specified") %>% 
  filter(nutrient == "zn_mg") %>% 
  mutate(nutrient = str_replace(nutrient, "zn_mg", "zinc")) %>% 
  filter(part != "tongues + cheeks") %>% 
  filter(part != "whole, no skin") %>% 
  group_by(nutrient, part) %>% 
  summarise_each(funs(mean, std.error), concentration) %>%
  group_by(nutrient) %>% 
  arrange(mean) %>% 
  ggplot(aes(x = reorder(part, mean), y = mean)) + geom_point(size = 3) +
  geom_errorbar(aes(ymin = mean - std.error, ymax = mean + std.error), width = 0.2) +
  facet_wrap( ~ nutrient, scales = "free") +
  geom_hline(yintercept = 11/10) +
  theme_bw() +
  theme(text = element_text(size=20),
        axis.text.x = element_text(angle=45, hjust=1)) + xlab("body part") + ylab("nutrient concentration (mg/100g edible portion)")




  microlement_body_part_plot <- grid.arrange(ca_plot, zn_plot, fe_plot, ncol = 3)
  ggsave("figures/microelements_body_part.png", width = 14, height = 8, plot = microlement_body_part_plot)
```
